requirements-dev.txt
-r requirements.txt
strawberry-graphql[debug-server]
black
